+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # YTech Changelog
 
+## []
+
+- Configurable temperature for smeltery and chimney
+
 ## [1.10.7]
 
 - Fixed crash in EMI workspace crafting widget rendering when empty layer